# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentacion API - Aulas",
        "description": "Listado de endpoints de producto - explicacion",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://apiaulas.test/"
        }
    ],
    "paths": {
...
        "/api/salida/{id}": {
            "get": {
                "tags": [
                    "Registro_salida"
                ],
                "summary": "Mostrar la información de una unica salida (venta)",
                "operationId": "c6689ff6ab964319e91b667ef442b6f6",
                "parameters": [
                    {
                        "name": "id_salida",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id_salida": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "unidades": {
                                            "type": "integer",
                                            "example": "5"
                                        },
                                        "id_factura_salida": {
                                            "type": "integer",
                                            "example": "8"
                                        },
                                        "id_producto": {
                                            "type": "integer",
                                            "example": "3"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Cliente] #id"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/usuario/{id}": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Mostrar la información de un solo usuario",
                "operationId": "e2b485ae0e81b04476da856d5bc824f8",
                "parameters": [
                    {
                        "name": "id_usuario",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id_usuario": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "example": "Andra Carmon"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "andrroman@gmail.com"
                                        },
                                        "contraseña": {
                                            "type": "integer",
                                            "example": "ASDAGFAFAGEGEW"
                                        },
                                        "rol_usuario": {
                                            "type": "string",
                                            "example": "2"
                                        },
                                        "tdoc_usuario": {
                                            "type": "integer",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Cliente] #id"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}